fix(sanitizers): clear ASan/UBSan baseline residue across analyzer/vm/coro/aot/ffi
Task 213 independent sanitizer-baseline audit. Real memory-safety defects
fixed and verified under ASan+UBSan (focused suite clean), with zero
regression in the non-sanitizer full suite (293/293):

- analyzer: enum-layout use-after-free — retire layouts replaced by the
  post-monomorphization re-analysis until analyzer teardown, since XrType
  copies cache the raw layout pointer (xanalyzer.h/.c, xanalyzer_visitor.c)
- vm: native struct field read+write misaligned access via memcpy (xvm_helpers.c)
- analyzer: compile-time constant left-shift signed overflow -> unsigned fold
  (xtype_ref_resolve.c)
- coro: coroutine-heap teardown UAF — drain the Region block pool only after
  fixed-heap cleanup (xsystem_heap.c)
- coro: timer cancel-queue dangling node — detach the embedded wheel timer
  before freeing coro->ext (xcoro.c)
- coro: countdown-latch close UAF — drain waiters before reading the
  possibly-freed scheduler (xcountdown_latch.c)
- thread: sys.Thread OS-stack overflow (three crash signatures, one cause) —
  floor the native stack at XR_WORKER_STACK_BYTES like the worker pool
  (xthread_obj.c)
- aot: map heap-buffer-overflow — skip the tagged-map flags probe for the
  smaller boolmap layout (xrt_coll.h)
- aot/ffi: non-NUL-terminated byte buffer overrun in libffi strlen —
  NUL-terminate the escaping fixed-heap copy (xvm.c)
- build: scale sanitizer CTest timeouts x8 (ASan/UBSan run 5-9x slower)

Remaining baseline residue tracked as follow-up: timer cancel-queue UAF on
the scheduler-teardown path, a multi-worker AOT BUS write, and ASan-
incompatible freestanding/benchmark/sysmon cases (gating, not product bugs).

+++ src/frontend/analyzer/xtype_ref_resolve.c
@@ -693,7 +693,7 @@ static bool ct_eval_binary(XaAnalyzer *analyzer, const AstNode *expr, XrCtValue 
                 case AST_BINARY_LSHIFT:
                     if (l < 0 || r < 0 || r >= 63)
                         return ct_fail(err, "bit shift is out of range");
-                    out->as.int_val = l << r;
+                    out->as.int_val = (int64_t) ((uint64_t) l << r);
                     return true;
                 case AST_BINARY_RSHIFT:
                     if (r < 0 || r >= 63)
